Combine Ui methods that are unnecessarily split in two
Methods in Ui return parts of the same standard response and can be combined for better readability.

Methods in Ui that return parts of the same standard response are combined.

Only one method for each standard response makes code more readable and easier to use.
