+++ src/main/java/duke/Parser.java
@@ -14,6 +14,7 @@ import duke.command.DeleteCommand;
 import duke.command.FindCommand;
 import duke.command.ListCommand;
 import duke.command.MarkCommand;
+import duke.exception.DukeException;
 
 /**
  * Deals with making sense of user commands and inputs.
