+++ src/test/java/duke/ParserTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.time.LocalDate;
 
+import duke.exception.DukeException;
 import org.junit.jupiter.api.Test;
 
 import duke.command.AddCommand;
